feat: extract business logic from SettingModal to SettingModalContent inorder to use withErrorBoundary to wrap it with error boundary
